requirements.txt
langchain
langgraph
langsmith 
numpy
pandas
flask_login
flask_sqlalchemy
flask_wtf
wtforms
flask_bcrypt
bcrypt
python-dotenv
langchain-groq
langchain_tavily
requests
google-api-python-client 
google-auth-oauthlib 
gspread 
requests
PyMupdf
flask_dance
flask
flask_mail
oauth2client
secrets
uuid
sentence-transformers
torch
tf-keras